fix(peer): encrypt-curve25519: avoid nonce reuse vulnerability
Add an additional blake3 hash step on the one-time message ed25519 key for use
as the chacha20poly1305 cipher nonce.

Previously, an attacker could brute-force search for a message which, when
hashed with the target public key, would have the same first 24 bytes for the
message public key, thus causing a nonce reuse when the target encrypts.

This way, the attacker would have to brute-force a message which produces a
blake3 hash (of the entire public key) which has the same first 24 bytes XOR'd
with the remaining 8 bytes. This is significantly more difficult to compute.

Signed-off-by: Christian Stewart <christian@paral.in>