refactor: drop kcp and xbee
These have dubious security and it is not in scope of bifrost to maintain
bespoke crypto code. All crypto should be done in third-party audited libs.

Drop KCP as it requires blockcrypt and blockcompress.

We are consolidating around quic-go instead of KCP anyway.

Drop xbee as it required KCP.

We may explore adding an xbee transport again in future w/ quic-go.

Signed-off-by: Christian Stewart <christian@aperture.us>